+++ utils.py
@@ -116,8 +116,10 @@ async def run_script(script: str | None = None):
                                                 stdout=asyncio.subprocess.PIPE,
                                                 stderr=asyncio.subprocess.PIPE)
 
+
 ###############################################################################
 ####################### UNIVERSAL STREAM DELIVERY SERVICE #####################
+##################### WITH HELPERS! ###########################################
 
 
 def get_field_value(item: Any, field: str):
@@ -140,7 +142,9 @@ async def stream_delivery(
     data_stream: deque,
     sort: bool | None = False,
     key: str | None = None,
-    group: str | None = None,
+    group: bool | None = False,
+    outer_key: str | None = None,
+    inner_key: str | None = None,
     filter_field: str | None = None,
     filter_param: str | None = None,
     page: int | None = None,
@@ -149,6 +153,7 @@ async def stream_delivery(
     old_snapshot = []
     while True:
         await asyncio.sleep(1)
+        total_items = None
         new_snapshot = list(data_stream)
 
         # Sorting
@@ -166,54 +171,55 @@ async def stream_delivery(
                 if get_field_value(item, filter_field) == filter_param
             ]
 
-        # DRAGON [2025-05-09]: Grouping logic below is hardcoded for ps_stream only.
-        # For now it's fine, but this function should stay stream-agnostic.
-        # Later: replace with a grouping_fn param (Callable) to support pluggable grouping per stream.
-
+        # DRAGON [2025-05-10]: SLAYED. Grouping modularized and decoupled.
         # Grouping
         if group:
-            ps_grouping_snapshot = list(data_stream)
-            new_snapshot = await ps_stream_grouping(items=ps_grouping_snapshot)
+            grouping_snapshot = list(data_stream)
+            new_snapshot = group_by_keys(
+                items=grouping_snapshot, outer_key=outer_key, inner_key=inner_key)
 
-        # Currently pagination only occurs for visitor activity stream. However,
-        # an item count takes place for every SSE at the moment. This only needs
-        # to run when the delivery needs to be paginated. Fine living with it as-
-        # is for now, but eventually need to isolate the functionality behind a
-        # conditional.
-
-        total_items = len(new_snapshot)
         # Pagination
         if page is not None and limit is not None:
-            start = (page - 1) * limit
-            end = start + limit
-            new_snapshot = new_snapshot[start:end]
+            paginating_snapshot = new_snapshot
+            new_snapshot, total_items = paginate(
+                page=page, limit=limit, snapshot=paginating_snapshot)
 
         if new_snapshot != old_snapshot:
             # needed to determine page total for frontend
             is_dc = is_dataclass(new_snapshot[0]) if new_snapshot else False
             serialized = [
                 asdict(item) if is_dc else item for item in new_snapshot]
-            yield f"event: meta\ndata: {json.dumps({'total_items': total_items})}\n\n"
+            if total_items:
+                yield f"event: meta\ndata: {json.dumps({'total_items': total_items})}\n\n"
             yield f"data: {json.dumps(serialized)}\n\n"
             old_snapshot = new_snapshot
+
+
+def paginate(page: int, limit: int, snapshot: list) -> tuple:
+    total_items = len(snapshot)
+    start = (page - 1) * limit
+    end = start + limit
+    new_snapshot = snapshot[start:end]
+    return (new_snapshot, total_items)
+
+
+def group_by_keys(items: list, outer_key: str, inner_key: str) -> list:
+    return_dict = {}
+    for item in items:
+        group = item.get(outer_key)
+        group_value = item.get(inner_key)
+        if not return_dict.get(group):
+            return_dict.update({group: {}})
+        if not return_dict[group].get(group_value):
+            return_dict[group][group_value] = []
+        return_dict[group][group_value].append(item)
+    return [{group: data} for group, data in return_dict.items()]
+
 
 ###############################################################################
 #### SYSTEM PROCESS ###########################################################
 ################# METRICS #####################################################
 ####################### STREAM ################################################
-
-
-async def ps_stream_grouping(items: list) -> list:
-    return_dict = {}
-    for item in items:
-        user = item.get("user")
-        ppid = item.get("ppid")
-        if not return_dict.get(user):
-            return_dict.update({user: {}})
-        if not return_dict[user].get(ppid):
-            return_dict[user][ppid] = []
-        return_dict[user][ppid].append(item)
-    return [{user: data} for user, data in return_dict.items()]
 
 
 async def ps_stream(script: str | None = None):